+++ Tasks/CodeQuality/PSScriptAnalyzer.ps1
@@ -2,11 +2,13 @@
     .SYNOPSIS
     Validates PSScriptAnalyzer passes
 #>
-task -Name PSScriptAnalyzer -Jobs SetVariables, {
+Add-BuildTask -Name PSScriptAnalyzer -Jobs SetVariables, {
+    $scriptFiles = Get-ChildItem $BuildRoot -File -Recurse |
+        Where-Object {$_.Extension -in '.ps1', '.psd1', '.psm1'}
     $settingsPath = Join-Path $BuildRoot 'PSScriptAnalyzerSettings.psd1'
     $settingsSplat = if (Test-Path $settingsPath) { @{ Settings = $settingsPath } } else { @{} }
-    $scriptFailures = $script:moduleFolders |
-    ForEach-Object { Invoke-ScriptAnalyzer $_ @settingsSplat }
+    $scriptFailures = $scriptFiles |
+    ForEach-Object { Invoke-ScriptAnalyzer $_.FullName @settingsSplat }
     $failures = foreach ($failure in $scriptFailures) {
         (
             "$($failure.ScriptName):$($failure.Line) - " +