+++ Tasks/CodeQuality/Pester/Pester.ps1
@@ -1,0 +1,5 @@
+<#
+    .SYNOPSIS
+    Parent test to run tests and confirm code coverage
+#>
+Add-BuildTask -Name Pester -Jobs ?PesterUnit, ?PesterIntegration, ?CodeCoverage