+++ Tasks/Content/Content.ps1
@@ -1,0 +1,5 @@
+<#
+    .SYNOPSIS
+    Runs content validation
+#>
+Add-BuildTask -Name Content -Jobs ?JSON, ?YAML