+++ Tasks/ModuleConventions/PublicFunctions.ps1
@@ -2,7 +2,7 @@
     .SYNOPSIS
     Validates all public functions are declared in the PSD1
 #>
-task -Name PublicFunctions -Jobs {
+Add-BuildTask -Name PublicFunctions -Jobs {
     $failures = Get-ChildItem "$BuildRoot\Public" | ForEach-Object {
         if ($_.BaseName -notin $script:psd1.FunctionsToExport) {
             "$($_.BaseName) is not in FunctionsToExport"