+++ Tasks/Pipeline/Validate.ps1
@@ -2,10 +2,8 @@
     .SYNOPSIS
     Main parent task to run the validation pipeline
 #>
-task -Name Validate -Jobs SetVariables,
-    ?ModuleVersion,
-    ?Changelog,
-    ?JSON,
-    ?Meta,
-    ?PSScriptAnalyzer,
-    ?Pester
+Add-BuildTask -Name Validate -Jobs SetVariables,
+    ?CodeQuality,
+    ?ReleaseHygiene,
+    ?Content,
+    ?ModuleConventions