+++ Tests/Assets/PassingModule/Tests/Integration/Get-Success.Integration.Tests.ps1
@@ -1,6 +1,6 @@
 BeforeDiscovery {
     $moduleName = 'PassingModule'
-    Get-Module $moduleName | Remove-Module 
+    Get-Module $moduleName | Remove-Module
     Import-Module "$PSScriptRoot\..\..\$moduleName.psd1" -Force
 }